skillmate/ai/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
pydantic>=2.5.2
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.1

# Vector database and embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Document processing
pypdf>=3.17.1
unstructured>=0.10.30
pdf2image>=1.16.3
pillow>=10.1.0
pytesseract>=0.3.10

# OpenAI dependencies
openai>=1.3.7
tiktoken>=0.5.1

# Google Gemini dependencies
google-generativeai>=0.3.1
langchain-google-genai>=0.0.5

# Web interface
flask>=2.3.3
flask-cors>=4.0.0

# Utilities
colorama>=0.4.6
tabulate>=0.9.0 